Hello
hello
this site is impossible to navigate and you have provided absolutely zero way for me to contact you, so hopefully this will work
While I know that you guys are fans of EE and you are doing this for him, I suspect you guys realize that Biribiri has absolutely nothing to do with Storm.
None of his abilities fit with any of Misaka's. There's another lightning based hero who fits her almost perfectly, however, and that is Zeus.
I am incredibly sad that you guys are making this mod for Storm, and I understand the fact that EE doesn't play Zeus, plays a lot of Storm, and Storm is a much more competitively viable hero.
I also understand the fact that you guys are doing a lot of work, and that this is by no means easy. 
The reason why I am writing this is because I would like to tell you that me, and many others, would really appreciate if you ported the files for Zeus. While this would definitely be some work, I hope you understand that Zeus fits Misaka much more.
I really hope you will consider this, as (correct me if I'm wrong), the only extra work required would be some animations for Zeus's spells and some minor code rewriting, as the majority of the hard work you guys (lines, model, hero animations, sounds) could be simply copied and pasted.
Please listen to my humble request, because this is almost a matter of life and death for me. Biribiri is one of my favorite characters and I would hate to not be able to use your great work because it was done on a hero that she doesn't fit with.